+++ libs/core/src/lib/combobox/combobox.module.ts
@@ -1,4 +1,5 @@
 import { NgModule } from '@angular/core';
+import { ComboboxItemDirective } from './combobox-item.directive';
 import { ComboboxComponent } from './combobox.component';
 import { ListGroupPipe } from './list-group.pipe';
 
@@ -7,7 +8,7 @@ import { ListGroupPipe } from './list-group.pipe';
  * Use `ComboboxComponent` import instead.
  */
 @NgModule({
-    imports: [ComboboxComponent, ListGroupPipe],
-    exports: [ComboboxComponent, ListGroupPipe]
+    imports: [ComboboxComponent, ComboboxItemDirective, ListGroupPipe],
+    exports: [ComboboxComponent, ComboboxItemDirective, ListGroupPipe]
 })
 export class ComboboxModule {}