+++ libs/docs/core/calendar/e2e/calendar.e2e-spec.ts
@@ -261,6 +261,10 @@ describe('calendar test suite', () => {
             const today = new Date();
             const endDate = new Date(today);
             endDate.setDate(today.getDate() + 7);
+            const nextDayAfterEndDate = new Date(endDate);
+            nextDayAfterEndDate.setDate(nextDayAfterEndDate.getDate() + 1);
+
+            console.log(endDate.getDay());
 
             if (endDate.getDate() < today.getDate()) {
                 await click(nextMonthButton);
@@ -276,9 +280,7 @@ describe('calendar test suite', () => {
             expect(
                 await (
                     await $(
-                        `fd-calendar-special-day-example .fd-calendar__item[data-fd-calendar-date-day="${
-                            endDate.getDate() + 1
-                        }"]:not(.fd-calendar__item--other)`
+                        `fd-calendar-special-day-example .fd-calendar__item[data-fd-calendar-date-day="${nextDayAfterEndDate.getDay()}"]:not(.fd-calendar__item--other)`
                     )
                 ).getAttribute('class')
             ).not.toContain('-legend-');