+++ libs/docs/core/combobox/examples/combobox-template-example.component.ts
@@ -1,17 +1,23 @@
 import { JsonPipe } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { ComboboxComponent } from '@fundamental-ngx/core/combobox';
+import { SearchHighlightPipe } from '@fundamental-ngx/cdk/utils';
+import { ComboboxComponent, ComboboxItemDirective } from '@fundamental-ngx/core/combobox';
 import { ListModule } from '@fundamental-ngx/core/list';
+
+interface ComboboxItem {
+    name: string;
+    icon: string;
+}
 
 @Component({
     selector: 'fd-combobox-template-example',
     templateUrl: './combobox-template-example.component.html',
     standalone: true,
-    imports: [ComboboxComponent, FormsModule, ListModule, JsonPipe]
+    imports: [ComboboxComponent, ComboboxItemDirective, FormsModule, ListModule, JsonPipe, SearchHighlightPipe]
 })
 export class ComboboxTemplateExampleComponent {
-    values = [
+    values: ComboboxItem[] = [
         { name: 'Photo Voltaic', icon: 'photo-voltaic' },
         { name: 'Settings', icon: 'settings' },
         { name: 'Heating Cooling', icon: 'heating-cooling' },
@@ -23,6 +29,8 @@ export class ComboboxTemplateExampleComponent {
         { name: 'Shield', icon: 'shield' },
         { name: 'Journey Change', icon: 'journey-change' }
     ];
+
+    model: ComboboxItem;
 
     selected: any;